move handleGuestLogout to hook
